requirements.txt
Babel==2.6.0
Django==2.1
django-ckeditor==5.6.1
django-jalali==3.1.0
django-js-asset==1.1.0
django-phonenumber-field==2.0.1
jdatetime==3.0.2
phonenumbers==8.9.12
Pillow==5.2.0
pluggy==0.7.1
py==1.6.0
pytz==2018.5
six==1.11.0
tox==3.2.1
virtualenv==16.0.0
